Day9/Agent-4-Experience-Retriever-Agent/requirements.txt
streamlit
langchain
google-generativeai
faiss-cpu
langchain-google-genai